+++ internal/gen/process/process.pb.go
@@ -1722,7 +1722,7 @@ const file_tool_process_process_proto_rawDesc = "" +
 	"\vStreamInput\x12\x1b.process.StreamInputRequest\x1a\x1c.process.StreamInputResponse(\x01\x12B\n" +
 	"\tSendInput\x12\x19.process.SendInputRequest\x1a\x1a.process.SendInputResponse\x12E\n" +
 	"\n" +
-	"SendSignal\x12\x1a.process.SendSignalRequest\x1a\x1b.process.SendSignalResponseB;Z9github.com/TencentCloudAgentRuntime/ags-go-sdk/pb/processb\x06proto3"
+	"SendSignal\x12\x1a.process.SendSignalRequest\x1a\x1b.process.SendSignalResponseBEZCgithub.com/TencentCloudAgentRuntime/ags-go-sdk/internal/gen/processb\x06proto3"
 
 var (
 	file_tool_process_process_proto_rawDescOnce sync.Once